producer/requirements.txt
gdown==5.2.0
matplotlib==3.8.2
msgpack_python==0.5.6
numpy==2.0.1
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
pika==1.3.2
PyQt5==5.15.11
PyQt5_sip==12.13.0
torch==2.5.0.dev20240726+cu121
ultralytics==8.2.28
psycopg2-binary